requirements.txt
Django>=4.2,<5.2
djangorestframework>=3.14
djangorestframework-simplejwt>=2.9
psycopg2-binary>=2.9
django-filter>=23.1
pytest>=8.0
pytest-django>=4.5
factory-boy>=3.2
python-dotenv>=1.0
gunicorn>=20.1
coverage>=7.0
pytest-cov>=4.0
